Digital elevation model (DEM) of the Mongón study area.

A raster::raster() object (EPSG:32717) representing altitude (ASTER GDEM, LP DAAC 2012). For more details, please refer to Muenchow et al. (2013).

Format

A raster::raster() with 117 rows and 117 columns:

dem

Altitude in m asl.

References

Muenchow, J., Bräuning, A., Rodríguez, E.F. & von Wehrden, H. (2013): Predictive mapping of species richness and plant species' distributions of a Peruvian fog oasis along an altitudinal gradient. Biotropica 45, 5, 557-566, doi: 10.1111/btp.12049.

LP DAAC (2012): Land Processes Distributed Active Archive Center, located at the U.S. Geological Survey (USGS) Earth Resources Observation and Science (EROS) Center. Available at: https://lpdaac.usgs.gov/ (last accessed 25 January 2012).